getpythonregpath():  Squash compiler warning about mixing signed and unsigned types in comparison. Relatedly, `dataSize` is declared as DWORD, not as int, so change relevant cast from (int) to (DWORD).

+++ PC/getpathp.c
@@ -306,8 +306,8 @@ getpythonregpath(HKEY keyBase, int skipcore)
 				Py_ssize_t len = _tcslen(ppPaths[index]);
 				_tcsncpy(szCur, ppPaths[index], len);
 				szCur += len;
-				assert(dataSize > len);
-				dataSize -= (int)len;
+				assert(dataSize > (DWORD)len);
+				dataSize -= (DWORD)len;
 			}
 		}
 		if (skipcore)
